These have to be cast as int or else they implicitly get pushed as unsigned ints.
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@3593 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ drivers/cam/dc1394_cam_server/src/dc1394_cam_server/dc1394_cam_server.cpp
@@ -92,8 +92,8 @@ public:
     uint32_t min;
     uint32_t max;
     cd.cam->getFeatureBoundaries(feature, min, max);
-    set_param(cd.name + string("/") + param_name + string("_min"), min);
-    set_param(cd.name + string("/") + param_name + string("_max"), max);
+    set_param(cd.name + string("/") + param_name + string("_min"), (int)(min));
+    set_param(cd.name + string("/") + param_name + string("_max"), (int)(max));
 
     string p = cd.name + string("/") + param_name;
     if (has_param(p))
